Account for leap years in core account age
By converting the diff to a date, the date since 1970 is calculated. The
year of that date minus 1970 would then equal the number of years that
have passed, taking leap years into account.

The "Member" achievement (number of years) has been adjusted to use the
years calculated from the core.