Add 12 months if month diff is negative
If the month is negative, this means that a year has passed, *but not a
full year*. E.g. 2021/10/01 - 2020/11/01. While the month count is
offset by `years * 12`, this would still count as 0 years passing. By
adding an additional year when a year has passed, calculated by when the
current month is less than the created month, an additional 12 months is
added, that the negative month diff can then "count back" from.

Fixes #556